_change_priority_refactor_genotype
# PandaHelper
Added PandaHelper that loads the dataset and metrices to allow for quick access. For example you can find the dataset labels by feature. Get the attributes that are found in a specific set of dataset labels. It does check if a file has changed by checking the last format time and if that is after the last updated, it updates the file.

It is not intendend to remain there, but it is useful to make the work and the requests are blazing fast once the mapper is build.